requirements.txt
flet==0.24.1
sqlalchemy==2.0.35
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
pytz==2024.1
